Use the correct Pinpoint client for voice message (#7)
Confusingly, there are 2 Pinpoint clients. `Pinpoint` which works for voice and not for SMS and `PinpointSMSVoice` which works for voice but not SMS.

We were trying to use the `Pinpoint` class for sending voice messages which was not working. This changes up the code to use the correct client and things appear to be working now.