LG-2198 Tear down old code: (#16)
* LG-2198 Tear down old code:

There are a number of changes in this PR:

Remove the longcode adapter:

This was a fallback in case there were issues with the Pinpoint adapter. Now that that is stable it is no longer necessary

Remove Twilio Verify:

We used Twilio Verify for international SMS. The API we used has been deprecated and we are no longer using it so I pulled this code out

Use a custom STS Client:

We want to use assume the role for sending Pinpont messages in the same region that we send the messages. This commit uses a custom STS client to make that possible.