Pinpoint OTP timeout error (LG-3925) (#35)
To avoid raising exceptions in the IDP during OTP requests, we need to rescue Seahorse::Client::NetworkingError during calls to Pinpoint for SMS/Voice.

Voice already had a rescue case, so the SMS rescue was modeled after it. The multiple configs should be tried even if one of them raises an exception since the rescues are implemented within the loop.